Remove the db dir from shared dir and stop incorrectly mounting a mysql docker entrypoint script
The mysql docker container allows you to prive startup scripts for the database in this entrypoint directory. We were nmounting a directory named like a file, and it confuses the docker startup for mysql foring a restart on startup. This restart is now problematic since we are telling docker-compose to wait for mysql health before starting services that depend on it. Since we aren't using the startup script mechanism at all, simply removing it seems like cleanest strategy