Improve revised integration tests for local use by devs (#17838)
* Improve revised integration tests for local use by devs

The goal of this work is to make running the revised ITs locally easier for devs while not impacting the robustness of ITs running in Github actions CI pipeline

* Internalize dependency startup conditional within template.py

Having dependency-service startup conditional setup use a centralized lookup with a sane default simplifies individual IT cases and will drive consistency across all existing and future IT cases

* Remove the db dir from shared dir and stop incorrectly mounting a mysql docker entrypoint script

The mysql docker container allows you to prive startup scripts for the database in this entrypoint directory. We were nmounting a directory named like a file, and it confuses the docker startup for mysql foring a restart on startup. This restart is now problematic since we are telling docker-compose to wait for mysql health before starting services that depend on it. Since we aren't using the startup script mechanism at all, simply removing it seems like cleanest strategy

* Clean up defunct comment

This comment has lost meaning after previous change to remove the init dir db for mysql as it is not needed