[WIP] Split up iq files and make tests run on all 3 engines
This PR splits up the iq files into smaller files at the testcase level. The files are grouped in subdirectories at the testsuite level.

This version also makes the tests run for all 3 engines.  The file name of an iq file indicicates what engine(s) the test uses:

If all 3 engines expect the same results:
*.all.iq: AllDruidEnginesComponentSupplier

If one of the enigines expect different results:
*.std.iq: StandardComponentSupplier
*.dart.iq: DartComponentSupplier
*.msq.iq: StandardMSQComponentSupplier