Kafka: Emit additional consumer metrics.
Adds more metrics to the ones originally added in #14582. These metrics
help provide insight into the workings of the Kafka consumer.

As a bonus, one of the metrics, "kafka/consumer/recordsLag", has a
dimension "partition" which can be used to more easily determine which
tasks are reading which partitions. It also provides a view of lag from
the task perspective rather than the Overlord perspective, which is
useful if the Overlord is ever unable to report metrics properly.