k8s: Update fabric8 to 6.13.1, switch to Vert.x http client. (#17913)
* k8s: Update fabric8 to 6.13.1, switch to Vert.x http client.

fabric8 6.x uses the OkHttp client by default, which itself uses an
unbounded-size cached pool, with 60s timeout, for making requests.
In production with large task counts, this leads to lots of thread churn.

The Vert.x client is the default in fabric8 7.x, and is better behaved.
By default it uses a worker pool of 20 threads (see javadoc for VertxOptions).

This patch uses fabric8 6.13.1 rather than some later version, because
higher versions are not compatible with Jackson 2.12.

* Fix call.

* Update licenses.yaml. Override thread pool configs.

* Remove annotation.

* Fix style.

* Update licenses.yaml.

* Add dep.