Fix filters on TIMESTAMP_TO_MILLIS over TIME_FLOOR. (#17871)
* Fix filters on TIMESTAMP_TO_MILLIS over TIME_FLOOR.

It is possible for toQueryGranularity to return nonnull for BIGINT rhs,
in the case where the time floor expression is wrapped in
TIMESTAMP_TO_MILLIS. Prior to this patch, that would throw an (incorrect)
error about expecting a literal.

This patch fixes the defensive error to have the correct message
(in Calcites) and also fixes the filter translation code to not hit that
defensive check (in Expressions).

* Fix error message.