Preserve column names as-is for unnest queries in the decoupled planner mode.
Unnest queries in the decoupled planner mode adds a suffix "_null" for some
cases. Currently the PlannerResult is based on the built query's output type rather
than the planned output type (which shouldn't contain the "_null" suffix). This is the
same for "coupled" planner.

The planner result is used for populating things in the sql direct statement. The sql row
transformer depends on the planner result, so this PR fixes that by aligning with the planned
output type.