Fix window function quidem tests (#18007)
When there are not enough order by columns to fix the order of a window functions, occassionally the query result may change.  This is especially true for order sensetive window functions such as first_value, last_value, or row_number.  This PR fixes the tests by adding more columns to the window order by list to fix the order.