Fix incorrect FixedBucketsHistogram counts when merging indexes (#18049)
* Fix incorrect FixedBucketsHistogram counts under certain index merging conditions

When `RowCombiningTimeAndDimsIterator.moveToNext() is merging indexes, if `MergingRowIterator.hasTimeAndDimsChangedSinceMark()` is `false`, then `AggregateCombiner.reset()` will be called through the call sequence `moveToNext()` -> `combineToCurrentTimeAndDims()` -> `resetCombinedMetrics()`, if `soleCurrentPointSourceOriginalIteratorIndex >= 0`. When exactly is all that the case? I'm not sure. But the net result was occasional over-counting, often wildly so, since `reset()` wasn't resetting anything, just combining whatever it had with the reset value.

* Change getBytes() to getBytes(StandardCharsets.UTF_8) in unit test