Include datasource name in the failure messages.
When a segment publish error occurs, it's hard to tell which supervisor to
reset if there are multiple supervisors consuming from the same stream. This
change includes the datasource name to the error messages. Ideally we would also include
the supervisor ID in the error, but since the supervisor ID is currently the same as
the datasource name, this should suffice for now.