Update aws.sdk.version (#18068)
* Update aws.sdk.version

In order to support EKS Pod Identity credentials, the minimum sdk version should be 1.12.746.
Its compatible with the current API.
https://docs.aws.amazon.com/eks/latest/userguide/pod-id-minimum-sdk.html

* Update AWS SDK for Java version in licenses.yaml

This change updates the version of the AWS SDK for Java to ensure compatibility with the latest features and security improvements. Without this update, the application may miss out on critical updates and enhancements.

* Updated version from 1.12.638 to 1.12.784 to leverage new features and fixes.
* Maintained existing license information and module details for consistency.

Signed-off-by: Thales Ceolin <thales@ulevel.com>

---------

Signed-off-by: Thales Ceolin <thales@ulevel.com>