Ensure response context is flushed at end of results accumulation to fix partial results behavior (#18025)
When there's an intermittent segment load issue on a historical, queries that touch the segment return partial results (despite the default behavior of returnPartialResults: true). This issue specifically occurs when there are multiple query runners and the missing segment is at the end of the timeline; I tested a few scenarios where the missing segment is at the beginning or in the middle of the timeline, and in those cases, the query throws an missing segment exception as expected.

The problem was that the response context header was written eagerly in initialize() and skipped in flush() once it's already initialized. However, query runners may still add entries to the response context after initialization. As a result, when ReportTimelineMissingSegmentQueryRunner sets the response context with the missing segments, it doesn't get propagated in the actual response sent back to the upstream server due to the short-circuit logic in flush() and other places.

* Preserve the original comment re limiting the size

* Comment clarity

* Try adding an IT

* Update javadocs for IT

* Revert the fix temporarily to verify IT failure

* Rename for clarity

* Switch to twitterstream datasource for query retry test because it has multiple segments.

- The twitterstream datasource has multiple segments (3 to be exact) as we can see in
  query-retry-sample-data.sql and the other similar sqls.
- On the other hand wikipedia_editstream datasource only has 1 segment which is less
  useful esepcially to test scenarios where there are multiple segments missing.
- Update the count for this in test expectation.

* Revert "Revert the fix temporarily to verify IT failure"

This reverts commit 36842bfc058fb19bae2672b8aff55ca5204bc95b.

* Rename file to twitterstream

* Revert "Revert "Revert the fix temporarily to verify IT failure""

This reverts commit b873df084129ebb2b4bb951a6cda986cf35c7d89.

* Revert "Revert "Revert "Revert the fix temporarily to verify IT failure"""

This reverts commit 8cc2d3e73d2042b50273dc503a5b14fe15342220.

* Address review comments.

* Add only validateAndWriteHeader() & commentary for the test