Quidem: Improved handling for subdirectories in quidem.filter.
Reworks the handling of path traversal to apply filter patterns to
the entire relativized path, not just the filename.

This helps with nested test files like `qaWin/basics_window_funcs.01.all.iq`.
Previously, `-Dquidem.filter=qaWin/basics_window*` would not match
that file, but `-Dquidem.filter=basics_window*` would. Now, the prior
matches and the former does not. This is in line with the behavior of
star (*) wildcards in shells, where they apply to individual path parts.