SQL: Parse queries immediately for non-JDBC endpoints.
SET statements (#17894) provide a new way to set query context. However,
they are not available early enough to inform important operations like
engine selection and planner rule construction. This patch moves parsing
as early as possible: immediately on receiving the query from the user.

As a consequence of moving parsing prior to statement creation, with
this change, SQL queries that cannot be parsed are no longer logged.