Fix mark used API response serialization (#18140)
Bug:
Some APIs in `DataSourcesResource` and `OverlordDataSourcesResource` are not able to
serialize out the JSON response.

__Note that the API is handled correctly by the server, it just fails to send back an appropriate response.__

This bug was introduced a while back in #17545 
While most of the APIs already handled sending back a JSON response, some APIs didn't.

Fix:
Update all APIs in `DataSourcesResource` and `OverlordDataSourcesResource` to handle
sending back a JSON response by annotating with `@Produces(MediaType.APPLICATION_JSON)`