Dart: Wait for a merge buffer if none is available. (#18075)
* Dart: Wait for a merge buffer if none is available.

For Dart-only workloads, it is not possible for merge buffers to be
exhausted, because the number of concurrent queries is capped at the
number of merge buffers by DartWorkerMemoryManagementModule. However,
When native and Dart queries are run on the same Historical, it is
possible for buffers to be held by native queries. In this case, Dart
workers must wait for buffers to become available.

This waiting now happens in the worker runner thread, and lasts for the
configured query timeout. At this time, the waiting is not interrupted
when a cancellation request arrives from the controller. This should be
addressed in another patch.

* Add test coverage for DartProcessingBuffersProvider.

* Changes from review.

* Add tests.

* More tests.