Fix ordering for certain floats in row-based frames.
Row-based frames had an incorrect sort order for certain float and double
values. The problem was an incorrect transformation in TransformUtils from
natural form to bits-sortable form. The now-legacy transformation compares
the sign and exponents properly for negative numbers, but compares significands
backwards, so if two negative floats have the same exponent they sort backwards.
It is fixed by flipping all bits for negative numbers, instead of just the
sign bit and exponent bits.

The main change is in TransformUtils, where the following methods gain
a `FrameType` parameter: `transformFromDouble`, `detransformToDouble`,
`transformFromFloat`, and `detransformToFloat`. With frame type
`ROW_BASED_V1`, the legacy transformation is used; with `ROW_BASED_V2`
the correct transformation is used.

All other changes are to support rolling updates. Because older code
and newer code may coexist during a job, a context parameter
"rowBasedFrameType" (default V1) is added. It's set to V2 on the Broker,
because the Broker is updated after all data servers. By the time the
Broker updates, it is safe to use the new type. Users do not need to set
this property directly, nor should they.