Testcontainers for embedded tests.
This patch adds the following:

1) TestcontainerResource that wraps testcontainers.

2) MariaDBMetadataResource that adds a MariaDB metadata store, and a test
   EmbeddedMariaDBMetadataStoreTest that uses it.

3) Rename EmbeddedKafkaServer to KafkaResource, and switch it to use a
   Kafka 4.0 container instead of an in-process Kafka.

To simplify the code a bit, this patch also adds a configureCluster method
to EmbeddedResource, which can be used by resources to configure a cluster
after they are started. It avoids needing to pass the cluster to the resource's
constructor, which would make instantiation awkward.