Embedded tests for MSQ realtime queries.
This patch adds an `EmbeddedMSQRealtimeQueryTest`, and `EmbeddedMSQApis` to
support it.

This patch also moves various usages of RuntimeInfo to regular injection
rather than using a static function, and creates a RuntimeInfoModule that is
part of the embedded tests and delivers a normal, non-statically-injected
RuntimeInfo.

This was necessary because the MSQ tasks, as they start up within the
multi-task Indexer, need to see more than 100 MB of total Indexer memory
in order to believe that there is sufficient memory to move forwards.
Static injection of RuntimeInfo makes this difficult, since the various
servers end up clobbering each other, and it isn't guaranteed that the
correct RuntimeInfo will be used.