Dart: Async queries to realtime servers.
Prior to this patch, queries from MSQ workers to data servers would
be initiated in the processing pool, and would block the processing
pool until results started coming in. This patch addresses it with
the strategy:

1) Update DataServerClient to return a future that resolves when the
   response starts being written.

2) Split DataServerQueryHandler into DartDataServerQueryHandler and
   IndexerDataServerQueryHandler. The Dart version doesn't do retries
   and doesn't follow segments to other data servers. It just returns
   the async future from DataServerClient. The Indexer (tasks) version
   retains the prior logic and isn't really async. I didn't attempt to
   asyncify its retry logic in this patch.

3) Add ReturnOrAwait.awaitAllFutures, which allows processors to wait
   for a future to resolve.

4) Update ScanQueryFrameProcessor and GroupByPreShuffleFrameProcessor
   to give up the processing thread when waiting for a data server query
   to come back.

Additionally, to simplify DataServerClient, cancellations are now
issued without using a scheduled executor. There should be no need for
this, because the service client is async.