Fix supervisor to tasks view naviagation in the console (#18298) (#18306)
With apache/druid#18082, the supervisor ID is no longer the same as the datasource name, which breaks the navigation from the supervisors page to the corresponding tasks view. To fix the broken filter during navigation, use the groupId filter instead of datasource name when navigating from the supervisors to tasks view.

Co-authored-by: Abhishek Radhakrishnan <abhishek.rb19@gmail.com>