NilStageOutputReader: Fix interruptible static initializer. (#18336)
The "write" method of "Channels.newChannel(baos)" is interruptible, even
though "baos" is a local nonblocking stream. This means the static
initializer of NilStageOutputReader is also interruptible, which is bad
because it might get interrupted if it's running in a processing thread.
This leads to the class being unusable with a ExceptionInInitializerError.

The fix is to hard-code the empty frame file. The new static has the same
content as the old one.