Add utility method to EmbeddedClusterApis to run tasks (#18349)
Description:
Embedded tests currently run into serialization issues when submitting a `Task` or
`CompactionSupervisorSpec` payload to the cluster using any `OverlordClient`.
This can happen because the `ObjectMapper` instance used by the `OverlordClient` might not have the
necessary Jackson modules loaded.
The hack so far has been to ensure that the first server in an embedded cluster is always an Overlord.
But this can quickly run into issues, e.g. when running a cluster with basic auth enabled where a
Coordinator must be started before all other servers.

Fix:
- Fix up methods `EmbeddedClusterApis`: `leaderOverlord`, `leaderCoordinator` and `anyBroker` to use
the appropriate servers. The request is still sent over the wire even if the client bound on a server is used
to call itself.
- Add utility methods in `EmbeddedClusterApis` to `submitTask`, `runTask`, `postSupervisor`