Fix ColumnarFrameCursorFactory canVectorize for virtual columns, aggregators. (#18385)
Previously, ColumnarFrameCursorFactory's canVectorize method didn'tc
consider virtual columns or aggregators from the CursorBuildSpec. This
meant that it would potentially return true for situations where a
virtual column or aggregator really couldn't vectorize, leading to failures
later on when the query engine attempts to create a vector cursor.

The change in RowFrameCursorFactory doesn't change any functionality,
it just structures the code more clearly by moving the CursorHolder to
an inner class instead of an anonymous class.