SQL: Update CaseToCoalesceRule to ignore irrelevant casts.
Previously, CaseToCoalesceRule ignored nullability casts when deciding
if "x" and "y" in "CASE WHEN x IS NOT NULL THEN y" match and could
therefore form the basis of a COALESCE. This was too strict, since it
considered a cast from VARCHAR to VARCHAR(2000), which can happen after
applying JSON_VALUE, to be relevant. It is not relevant, because it does
not affect how we execute the query.

This patch updates the logic to ignore nullability casts and also casts
that change irrelevant properties of the type. ("Irrelevant" being defined
as "properties that do not affect how the query executes".)