Decouple AllTrue, AllUnknown BitmapColumnIndex from ColumnIndexSelector.
There is no real need for these objects to require a ColumnIndexSelector
for construction. All they need is the bitmap factory and row count. This
change makes it easier to use these classes from places where a
ColumnIndexSelector is not available, such as implementations of
semantic indexes.

This also brings these classes more in line with AllFalseBitmapColumnIndex,
which just takes a bitmap factory.