Ensure DruidMeta is closed properly.
DruidMeta owns an executor service, which spawns a thread that retains
a reference to it. This service must be shut down for the DruidMeta to
be garbage-collectable, which is important when DruidAvaticaTestDriver
is closed. Otherwise, any segments retained by the DruidMeta will stick
around.

This fix includes a `@LifecycleStop` so the same shutdown happens when
an actual server is shutting down. It seems like good hygiene.