Support truncation of SQL index names in metadata store
For metadata stores with upper limits on the base table name, creation of indices can sometimes fail.
Because these failures are not critical to the operation of the cluster, they are logged as warnings and can result in performance degradatation.
This patch allows for index names to be truncated according to an engine-specific limit, so that they are always created.