Bouncer: Acquire child tickets before parent tickets.
Follow up to #18520. The logic introduced in that patch for
parent/child relationships was not ideal, since it would lead
to parent tickets being acquired even if the child was at its
max count. In production, this would cause a query with
maxThreads = 1 to acquire more than 1 thread ticket from the
global Bouncer.