Remove warn log for uknown task location in KubernetesPeonLifecycle (#18574)
This log can become noisy and misleading when running a lot of MSQ tasks with K8s task runner.
The controller starts checking worker status quickly once the worker task has started but not got
to a running state yet. This results in a large volume of logs that can mislead operators and/or users
about the health of their indexing service.