Implement flushPeriod to expire metrics past TTL for prometheus-emitter (#18598)
Partially addresses #14638. This implementation adds flushPeriod to the exporter strategy in the prometheus-emitter, which when configured, will stop emitting the metric if it has not been updated within the TTL. This does not account for scenarios where the metric is being emitted but with different label values. This PR can be enhanced in the future to track all label combinations.