Move all revised ITs to embedded tests (#18622)
Description
------------
This patch moves all ITs from `integration-tests-ex` to `embedded-tests`.
Most of the files have just been moved to the embedded tests module.
The moved tests are all functional but they still retain the IT flavor and might thus be a little verbose.
This has been done to keep the diff small and simplify reviews.
Later, the tests can be refactored to be more concise and "embedded test"-like.

Changes
---------
- Add `AzureStorageResource` which uses
[`AzuriteContainer`](https://testcontainers.com/modules/azurite/) for Azure storage tests
- Add utility method `AzureAccountConfig.createStorageEndpointUrl()`
- Use the `druid.azure.protocol` property to determine if the endpoint URL points to a test Azurite container
- Add `GoogleCloudStorageResource` which uses
[`FakeGcsServerContainer`](https://testcontainers.com/modules/fake-gcs-server/) for GCS tests
- Add test extension module `GoogleStorageTestModule` to bind `Storage` to a test instance with no credentials
- Use `MinIOStorageResource` for S3 tests
- Remove old IT groups `batch-index`, `input-source`, `input-format` (details below)
- Remove all `integration-tests-ex` code and GHA workflows