Fix query status metric reporting bug in router (#18639)
Currently, all query errors that did not originate from the router are incorrectly being recorded as successful in metric emission at the router. For example, a query that timed out on a broker would be reported as a successful query by the router during metric emission.

This is because the current code checks an exception on the response object, rather than the actual status code of the response. This is incorrect because the router is operating as a proxy, and a "successful" result is considered one which was proxied + received a response, and will not throw an exception. We should be checking the status code as well, since this is the true status determined by the downstream broker.