Add compactTask/busySlot/count metric to compact duty.
This adds the compactTask/busySlot/count metric directly, rather than indirectly deducing it through the
formula for available task slots and maximum task slots and taking their minimum.

The formula to compute available compact task slots in a run is:
availableCompactionTaskSlots = Math.max(0, compactionTaskCapacity - busyCompactionTaskSlots);

Often times, it's confusing why compactTask/availableSlot/count is lower than expected.
It turns out that the compact duty using the native engine just caps it using maxNumConcurrentSubTasks,
regardless of the phase a current compact supervisor is in. This is likely the safest and conservative thing to do.

This metric should help operators better plan for compaction task slots in a MM-based setup.