Fix extra router statusCode metric edge cases (#18699)
The `statusCode` dimension on `query/time` metric will misreport if connection exceptions occur while router/broker are communicating (e.g. there's no HTTP status code yet – Jetty sets defaults this to 0) or when a client manually closes the connection while a downstream query is still in-flight. We cannot perfectly map every error in this case (nor do I think we should), but it's better than misreporting incorrect codes. Since we need to report a metric event anyways, maintaining a fallback of 500/200 if we cannot determine the error code is consistent with how things are handled already on the .*QueryResource classes and router should do the same.