Add dataSource and taskId dimensions to GroupByStatsMonitor for peons (#18709)
The GroupByStatsMonitor includes dataSource and taskId dimensions for metrics emitted on the peons.

Monitors on peons that previously emitted the id dimension from JettyMonitor, OshiSysMonitor, JvmMonitor, JvmCpuMonitor, JvmThreadsMonitor and SysMonitor to represent the task ID are deprecated and will be removed in a future release. Use the taskId dimension instead.