Add task/waiting/time metric (#18735)
Adds a metric `task/waiting/time` that measures the time it takes for a task to be placed onto the task runner for scheduling+running. This time encompasses initial persistent storage insertion, lock acquisition time, any task queue-related latencies, noisy-neighbor-related slowness, and any other delays between when a task is first added and when it is scheduled for running on the task runner. Useful for debugging issues related to task startup.