Remove various version overrides from module poms. (#18708)
* Remove various version overrides from module poms.

Many poms explicitly specified versions for dependencies and plugins
that have centrally-managed versions in the master pom's dependencyManagement
and pluginManagement section.

For the most part, the overrides were simply restating the centrally-managed
version, so they weren't causing an actual divergence.

The most substantive change is that hamcrest is upgraded from 1.3 to 2.2.
Version 2.2 was already being used by kubernetes-overlord-extensions.