Don't kill tasks while a supervisor is stopping.
Previously, if a supervisor was stopped while discovering tasks or
updating their status, it could end up trying to kill those tasks because
the callbacks on the requests to those tasks could fail as a result of
the stopping. There should be no reason for an actively-stopping supervisor
to kill a task, so this patch causes the shutdown functions to be no-ops
while a supervisor is stopping.

This patch also ensures that when a supervisor transitions into the STOPPING
state, that state takes priority over any other state.