Allow large partition numbers in VersionedIntervalTimeline.
The experimental segment locking feature (forceTimeChunkLock: false)
reserves the first 32768 partition numbers for the "root generation",
and then uses an "atomic update groups" scheme to replace root segment
ranges with new sets of segments during reindexing operations.

OvershadowableManager, which manages this atomic update scheme, imposes
a limit of 32768 segments per time chunk. Previously, this applied even
to people that are not using segment locking. In this patch, the class
is now only used when segment locking is actually in play, meaning that
the limit is not imposed under normal conditions.