Refactor: Rename DataSourceTaskIdHolder and extract load spec holder to its own class (#18732)
Refactor:

Remove named injections of DataSourceTaskIdHolder in favor of:
- TaskHolder interface: implemented by PeonTaskHolder for peon processes and by NoopTaskHolder for all other servers, where the methods return null.
- LoadSpecHolder interface: implemented by PeonLoadSpecHolder for peon processes and DefaultLoadSpecHolder for all other servers.

This also removes some hacks that were added to avoid cyclic dependencies during Guice init. A module DefaultServerHolderModule that binds the above modules as part of the core injector module for all servers except CliPeon. For CliPeon, the binding to PeonTaskHolder and PeonLoadSpecHolder is explicitly done in CliPeon itself.