Include JDK version in test-reports so it's easier to find
Currently we use one uber test-report across both JDK versions 17 and 21.
Some tests may be flaky or even have JDK specific errors that can make it
slightly difficult to identify the source of the error.

So this patch should help which JDK/job to look into in case of errors rather than
having to hunt down between the different versions.