Include JDK version in test-reports so it's easier to find (#18817)
Currently, we generate a single consolidated test-report across both JDK 17 and JDK 21. Some tests may fail or flake only on one JDK version, which makes it harder to identify the source of the error.

This patch generates a separate test report per JDK version, which will allow us to identify the source of a failure by checking the corresponding JDK-specific job.