MostFragmentedIntervalFirstPolicy: Implement equals. (#18820)
* MostFragmentedIntervalFirstPolicy: Implement equals.

Also hashCode and toString. This is required for correct behavior,
since the policy will not be saved to the database on edit if it
compares equal to the existing, and the base class BaseCandidateSearchPolicy
does have an equals method. So, subclasses such as this one must
override equals and hashCode.

* Remove unused import.