Add task dimensions to ServiceEmitter.
This patch adds task dimensions to the ServiceEmitter so different
monitors don't need to pass the holder to get the dimensions and append
it during every monitor call. This also ensures that all task metrics emitted
contain the task dimensions by default and are emitted in the same format.

The dimensions added include  taskId, dataSource, taskType, groupId and id.
id is added primarily for backward compatibility as some monitors were previously
emitting the id dimension rather than taskId. We can remove it in later release as it's
been deprecated already.