Add task dimensions to `ServiceEmitter` (#18876)
All task metrics are now emitted with the following dimensions: taskId, dataSource, taskType, groupId, and id (for backward compatibility; id will be removed in favor of the taskId dimension in a future release).

This is achieved by wiring in the TaskHolder and getting the task dimensions from it in the ServiceEmitter initialization, so individual monitors no longer need to pass the task holder to retrieve and append task dimensions on every monitor call. Some monitors were previously missing these dimensions. This change ensures that all task metrics are emitted with these dimensions by default and in a consistent format.